opt: move opttester to separate package
Moving OptTester to an `opttester` sub-package. This allows code that
opttester depends on to use the remaining test utils.

Release note: None
