+++ pkg/sql/opt/testutils/opttester/explore_trace.go
@@ -12,7 +12,7 @@
 // implied. See the License for the specific language governing
 // permissions and limitations under the License.
 
-package testutils
+package opttester
 
 import (
 	"github.com/cockroachdb/cockroach/pkg/sql/opt"
