+++ pkg/sql/opt/testutils/opttester/path_cache.go
@@ -12,7 +12,7 @@
 // implied. See the License for the specific language governing
 // permissions and limitations under the License.
 
-package testutils
+package opttester
 
 import (
 	"bytes"
